refactor: extract `argus-abi`, `argus-monitor`, argus-rhai`, etc.
