+++ crates/argus-abi/src/test_utils/mod.rs
@@ -1,11 +1,11 @@
-//! Test helpers for ABI-related functionality.
+//! ABI test helpers.
 
 use std::sync::Arc;
 
 use argus_core::persistence::traits::AppRepository;
 use argus_store::SqliteStateRepository;
 
-use crate::abi::{AbiRepository, AbiService};
+use crate::{AbiRepository, AbiService};
 
 /// A simple ABI JSON for testing purposes.
 pub fn erc20_abi_json() -> &'static str {
