+++ crates/argus-api/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2024"
 [dependencies]
 argus-core.workspace = true
 argus-engine.workspace = true
+argus-abi.workspace = true
+argus-monitor.workspace = true
 axum.workspace = true
 serde.workspace = true
 serde_json.workspace = true
