+++ crates/argus-api/src/http_server/mod.rs
@@ -12,7 +12,7 @@ use std::{net::SocketAddr, sync::Arc};
 use abi::{delete_abi, get_abi_by_name, list_abis, upload_abi};
 use actions::{create_action, delete_action, get_action_details, get_actions, update_action};
 use argus_core::{config::AppConfig, metrics::AppMetrics, persistence::traits::AppRepository};
-use argus_engine::monitor::{MonitorPersistenceValidator, MonitorValidator};
+use argus_monitor::{MonitorPersistenceValidator, MonitorValidator};
 use auth::auth;
 use axum::{
     Router, middleware,
