+++ crates/argus-core/src/test_utils/mod.rs
@@ -2,6 +2,7 @@ pub mod actions;
 pub mod block;
 pub mod log;
 pub mod monitor;
+pub mod monitor_match;
 pub mod receipt;
 pub mod transaction;
 
@@ -9,5 +10,6 @@ pub use actions::ActionBuilder;
 pub use block::BlockBuilder;
 pub use log::LogBuilder;
 pub use monitor::MonitorBuilder;
+pub use monitor_match::{create_test_monitor_match, create_test_tx_monitor_match_with_hash};
 pub use receipt::ReceiptBuilder;
 pub use transaction::TransactionBuilder;
