+++ crates/argus-engine/src/engine/mod.rs
@@ -7,4 +7,6 @@ pub mod block_processor;
 pub mod filtering;
 pub mod outbox_processor;
 
+pub use argus_abi as abi;
+pub use argus_monitor as monitor;
 pub use argus_rhai as rhai;
