+++ crates/argus-engine/src/lib.rs
@@ -2,9 +2,8 @@
 //!
 //! Provides the ABI decoding, monitoring, and block processing engines.
 
-pub mod abi;
-pub mod engine;
-pub mod monitor;
-
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils;
+pub mod alert_manager;
+pub mod block_ingestor;
+pub mod block_processor;
+pub mod filtering;
+pub mod outbox_processor;
