--- crates/argus-engine/src/test_utils/mod.rs
@@ -1,9 +1,0 @@
-//! Test utilities for the argus-engine crate.
-
-mod abi;
-mod monitor_manager;
-mod monitor_validator;
-
-pub use abi::{create_test_abi_service, erc20_abi_json};
-pub use monitor_manager::create_test_monitor_manager;
-pub use monitor_validator::create_monitor_validator;
