+++ crates/argus-monitor/src/lib.rs
@@ -9,3 +9,6 @@ pub use interest_registry::InterestRegistryBuilder;
 pub use manager::{ClassifiedMonitor, MonitorAssetState, MonitorCapabilities, MonitorManager};
 pub use persistence_validator::{MonitorPersistenceValidationError, MonitorPersistenceValidator};
 pub use validator::{MonitorValidationError, MonitorValidator};
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_utils;
