--- src/test_helpers/http_client.rs
@@ -1,13 +1,0 @@
-use std::sync::Arc;
-
-use argus_core::config::HttpRetryConfig;
-use argus_dispatch::http_client::create_retryable_http_client;
-use reqwest::Client;
-use reqwest_middleware::ClientWithMiddleware;
-
-/// Creates a default HTTP client with retry capabilities for testing purposes.
-pub fn create_test_http_client() -> Arc<ClientWithMiddleware> {
-    let retryable_client = create_retryable_http_client(&HttpRetryConfig::default(), Client::new());
-
-    Arc::new(retryable_client)
-}
