--- src/test_helpers/mod.rs
@@ -1,28 +1,0 @@
-//! A set of helpers for testing
-
-mod abi;
-mod block;
-mod http_client;
-mod log;
-mod monitor;
-mod monitor_manager;
-mod monitor_match;
-mod monitor_validator;
-mod provider;
-mod receipt;
-mod transaction;
-
-pub use abi::{create_test_abi_service, erc20_abi_json};
-pub use block::BlockBuilder;
-pub use http_client::create_test_http_client;
-pub use log::LogBuilder;
-pub use monitor::MonitorBuilder;
-pub use monitor_manager::create_test_monitor_manager;
-pub use monitor_match::{
-    create_test_log_monitor_match, create_test_monitor_match, create_test_monitor_match_custom,
-    create_test_monitor_match_with_call, create_test_tx_monitor_match,
-};
-pub use monitor_validator::create_monitor_validator;
-pub use provider::mock_provider;
-pub use receipt::ReceiptBuilder;
-pub use transaction::TransactionBuilder;
